+++ example/lib/main.dart
@@ -6,11 +6,10 @@ main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   await translator.init(
-    localeDefault: LocalizationDefaultType.device,
+    localeType: LocalizationDefaultType.device,
     languagesList: <String>['ar', 'en'],
-    assetsDirectory: 'assets/langs/',
-    apiKeyGoogle: '<Key>', // NOT YET TESTED
-  ); // intialize
+    assetsDirectory: 'assets/lang/',
+  );
 
   runApp(LocalizedApp(child: MyApp()));
 }
@@ -38,8 +37,7 @@ class Home extends StatelessWidget {
     return Scaffold(
       drawer: Drawer(),
       appBar: AppBar(
-        title: Text(translator.translate('appTitle')),
-        // centerTitle: true,
+        title: Text('appTitle'.tr()),
       ),
       body: Container(
         width: double.infinity,
@@ -49,20 +47,24 @@ class Home extends StatelessWidget {
           children: <Widget>[
             SizedBox(height: 50),
             Text(
-              translator.translate('textArea'),
+              'textArea'.tr(),
               textAlign: TextAlign.center,
               style: TextStyle(fontSize: 35),
             ),
-            OutlinedButton(
-              onPressed: () {
-                translator.setNewLanguage(
-                  context,
-                  newLanguage: translator.currentLanguage == 'ar' ? 'en' : 'ar',
-                  remember: true,
-                  restart: true,
+            Wrap(
+              children: translator.locals().map((i) {
+                return OutlinedButton(
+                  onPressed: () {
+                    translator.setNewLanguage(
+                      context,
+                      newLanguage: i.languageCode,
+                      remember: true,
+                      restart: true,
+                    );
+                  },
+                  child: Text(i.languageCode),
                 );
-              },
-              child: Text(translator.translate('buttonTitle')),
+              }).toList(),
             ),
           ],
         ),